+++ CHANGELOG.md
@@ -17,6 +17,7 @@ first, also the PHP version required is 7.1, see [#426].
 - Allow auth adapters to be chained (@glensc, #441)
 - Remove Legacy CLI application (@glensc, #453)
 - Move Login-Back-Off logic to MySQL adapter only (@glensc, #455)
+- Use `symfony/var-exporter` for writing config files (@glensc, #456)
 
 [3.6.0]: https://github.com/eventum/eventum/compare/v3.5.6...master
 [#426]: https://github.com/eventum/eventum/pull/426